prooflink/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# AI & ML
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
scikit-learn==1.3.2

# File Processing
python-magic==0.4.27
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Cloud Storage
boto3==1.29.7
google-cloud-storage==2.10.0
dropbox==11.36.2

# Email
python-email-validator==2.1.0
emails==0.6

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# Payment
stripe==7.7.0

# Monitoring & Logging
sentry-sdk==1.38.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2

# QR Code Generation
qrcode[pil]==7.4.2

# Hashing & Signatures
hashlib-additional==1.0.0
ecdsa==0.18.0

# Rate Limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0

# WebSocket
websockets==12.0

# Timezone
pytz==2023.3
python-dateutil==2.8.2

# UUID
shortuuid==1.0.11

# JSON Web Tokens
pyjwt==2.8.0

# File Type Detection
filetype==1.2.0

# Image Processing
imageio==2.33.0
scikit-image==0.22.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Excel Processing
pandas==2.1.4
xlsxwriter==3.1.9

# Markdown Processing
markdown==3.5.1
pymdown-extensions==10.5

# HTML Processing
beautifulsoup4==4.12.2
lxml==4.9.3

# URL Processing
validators==0.22.0
urllib3==2.1.0

# Compression
zstandard==0.22.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1